+++ src/array.rs
@@ -684,8 +684,10 @@ pub trait ArrayValue:
     const SYMBOL: char;
     /// An ID for the type
     const TYPE_ID: u8;
-    /// Get the fill value from the environment
-    fn get_fill(env: &Uiua) -> Result<Self, &'static str>;
+    /// Get the scalar fill value from the environment
+    fn get_scalar_fill(env: &Uiua) -> Result<Self, &'static str>;
+    /// Get the array fill value from the environment
+    fn get_array_fill(env: &Uiua) -> Result<Array<Self>, &'static str>;
     /// Hash the value
     fn array_hash<H: Hasher>(&self, hasher: &mut H);
     /// Get the proxy value
@@ -730,8 +732,11 @@ impl ArrayValue for f64 {
     const NAME: &'static str = "number";
     const SYMBOL: char = 'ℝ';
     const TYPE_ID: u8 = 0;
-    fn get_fill(env: &Uiua) -> Result<Self, &'static str> {
-        env.num_fill()
+    fn get_scalar_fill(env: &Uiua) -> Result<Self, &'static str> {
+        env.num_scalar_fill()
+    }
+    fn get_array_fill(env: &Uiua) -> Result<Array<Self>, &'static str> {
+        env.num_array_fill()
     }
     fn array_hash<H: Hasher>(&self, hasher: &mut H) {
         let v = if self.is_nan() {
@@ -752,8 +757,11 @@ impl ArrayValue for u8 {
     const NAME: &'static str = "number";
     const SYMBOL: char = 'ℝ';
     const TYPE_ID: u8 = 0;
-    fn get_fill(env: &Uiua) -> Result<Self, &'static str> {
-        env.byte_fill()
+    fn get_scalar_fill(env: &Uiua) -> Result<Self, &'static str> {
+        env.byte_scalar_fill()
+    }
+    fn get_array_fill(env: &Uiua) -> Result<Array<Self>, &'static str> {
+        env.byte_array_fill()
     }
     fn array_hash<H: Hasher>(&self, hasher: &mut H) {
         (*self as f64).to_bits().hash(hasher)
@@ -767,8 +775,11 @@ impl ArrayValue for char {
     const NAME: &'static str = "character";
     const SYMBOL: char = '@';
     const TYPE_ID: u8 = 2;
-    fn get_fill(env: &Uiua) -> Result<Self, &'static str> {
-        env.char_fill()
+    fn get_scalar_fill(env: &Uiua) -> Result<Self, &'static str> {
+        env.char_scalar_fill()
+    }
+    fn get_array_fill(env: &Uiua) -> Result<Array<Self>, &'static str> {
+        env.char_array_fill()
     }
     fn format_delims() -> (&'static str, &'static str) {
         ("", "")
@@ -798,8 +809,11 @@ impl ArrayValue for Boxed {
     const NAME: &'static str = "box";
     const SYMBOL: char = '□';
     const TYPE_ID: u8 = 3;
-    fn get_fill(env: &Uiua) -> Result<Self, &'static str> {
-        env.box_fill()
+    fn get_scalar_fill(env: &Uiua) -> Result<Self, &'static str> {
+        env.box_scalar_fill()
+    }
+    fn get_array_fill(env: &Uiua) -> Result<Array<Self>, &'static str> {
+        env.box_array_fill()
     }
     fn array_hash<H: Hasher>(&self, hasher: &mut H) {
         self.0.hash(hasher);
@@ -819,8 +833,11 @@ impl ArrayValue for Complex {
     const NAME: &'static str = "complex";
     const SYMBOL: char = 'ℂ';
     const TYPE_ID: u8 = 1;
-    fn get_fill(env: &Uiua) -> Result<Self, &'static str> {
-        env.complex_fill()
+    fn get_scalar_fill(env: &Uiua) -> Result<Self, &'static str> {
+        env.complex_scalar_fill()
+    }
+    fn get_array_fill(env: &Uiua) -> Result<Array<Self>, &'static str> {
+        env.complex_array_fill()
     }
     fn array_hash<H: Hasher>(&self, hasher: &mut H) {
         for n in [self.re, self.im] {